google_vision/requirements.txt
google-cloud-vision
pyfiglet
psycopg2